+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `IObservable` interface
+- `ObservableObject` class
+- `ObservableWatcher` component
+- Observables sample scene
 
 ## [0.5.0] - 2022-12-03